SUM Solve Warnings on Variable
^D Solve warnings when uses MSVC warning level `/W4`.
 : Adjust string length variable type to `size_t` (`unsigned long long`).
*? Rename language type codename of Simplified Chinese.
 : From `zh-Hans` to `zh-CN`.
 : Modify URL links.
^? Fix help file.
 : Modify wrong command names in file `task.dat`.
 : Not `WORK`.

+++ modules/basic.hpp
@@ -263,6 +263,6 @@ void WCH_Sleep(int _ms);
 void WCH_printlog(wstring _mode, wstring _info);
 void WCH_read();
 bool WCH_save_func();
-int WCH_GetNumDigits(int _n);
+size_t WCH_GetNumDigits(size_t _n);
 
 #endif
